Using JteView instead of a resolved MvcView
We have migrated to static jte templates, which do not require us to work with untyped models and settings string params to model in an error-prone way.

For it to work, we've configured HttpMessageConverter for Jte, that renders generated JteModel's to output stream.

Also replaced @Controller with @RestController - we do not need to resolve views by path anymore, but we have to put the "produces" argument on our mappings to always return text/html